+++ PluginsAndFeatures/azure-toolkit-for-rider/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Changed
 - Update tool window icon
+- Reimplement Azurite support
 
 ## [4.0.0-preview.1] - 2024-01-24